+++ README.md
@@ -1,40 +1,34 @@
-This is a [Next.js](https://nextjs.org/) project bootstrapped with [`create-next-app`](https://github.com/vercel/next.js/tree/canary/packages/create-next-app).
+## Description
+
+Welcome to my Portfolio...
+
+-   [Next.js](https://nextjs.org/)
+-   [Mui](https://mui.com/)
+-   [ExpressJS](https://expressjs.com/fr/)
+-   [mongoDB](https://www.mongodb.com/fr-fr)
+-   [react-pdf](https://react-pdf.org/)
 
 ## Getting Started
 
-First, run the development server:
+Start backend:
 
 ```bash
-npm run dev
+cd backend
+
+yarn start
 # or
-yarn dev
-# or
-pnpm dev
-# or
-bun dev
+nodemon
 ```
 
-Open [http://localhost:3000](http://localhost:3000) with your browser to see the result.
+Run by default on [http://localhost:3000] (http://localhost:3000)
 
-You can start editing the page by modifying `pages/index.js`. The page auto-updates as you edit the file.
+Start frontend:
 
-[API routes](https://nextjs.org/docs/api-routes/introduction) can be accessed on [http://localhost:3000/api/hello](http://localhost:3000/api/hello). This endpoint can be edited in `pages/api/hello.js`.
+```bash
+cd frontend
 
-The `pages/api` directory is mapped to `/api/*`. Files in this directory are treated as [API routes](https://nextjs.org/docs/api-routes/introduction) instead of React pages.
+yarn dev
+```
 
-This project uses [`next/font`](https://nextjs.org/docs/basic-features/font-optimization) to automatically optimize and load Inter, a custom Google Font.
-
-## Learn More
-
-To learn more about Next.js, take a look at the following resources:
-
-- [Next.js Documentation](https://nextjs.org/docs) - learn about Next.js features and API.
-- [Learn Next.js](https://nextjs.org/learn) - an interactive Next.js tutorial.
-
-You can check out [the Next.js GitHub repository](https://github.com/vercel/next.js/) - your feedback and contributions are welcome!
-
-## Deploy on Vercel
-
-The easiest way to deploy your Next.js app is to use the [Vercel Platform](https://vercel.com/new?utm_medium=default-template&filter=next.js&utm_source=create-next-app&utm_campaign=create-next-app-readme) from the creators of Next.js.
-
-Check out our [Next.js deployment documentation](https://nextjs.org/docs/deployment) for more details.
+Open [http://localhost:3000](http://localhost:3000) with your browser
+If backend is already running on `:3000` switch to `:3001`:3001
