+++ frontend/package.json
@@ -20,8 +20,9 @@
     "@fortawesome/react-fontawesome": "^0.2.0",
     "@mui/icons-material": "^5.15.15",
     "@mui/material": "^5.15.15",
-    "@react-pdf/renderer": "^3.1.15",
+    "@react-pdf/renderer": "^3.4.4",
     "@reduxjs/toolkit": "^2.0.1",
+    "file-saver": "^2.0.5",
     "next": "14.0.4",
     "react": "^18",
     "react-copy-to-clipboard": "^5.1.0",
